pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "full-quantum-deep-learning"
version = "0.1.0"
description = "Full Quantum Deep Learning (QDL) reference implementation with PennyLane + PyTorch."
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
authors = [
    {name="Your Name", email="your.email@example.com"}
]
dependencies = [
    "pennylane>=0.36",
    "pennylane-lightning>=0.36",
    "torch>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "rich>=13.7",
    "tyro>=0.6.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.4.8",
    "mypy>=1.10",
    "black>=24.4"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "N"]
ignore = ["E501"]